积极快乐健康了

+++ src/main/java/com/test/day03/homework/Shape.java
@@ -4,6 +4,6 @@ package com.test.day03.homework;
  * Created by JiChenGuang on 16/3/23.
  */
 public abstract class Shape {
-        public abstract int getArea();
+        public abstract double getArea();
 
 }
